requirements.txt
# Aria v1.0 — Personal AI Desktop Assistant
# ============================================
# Install: pip install -r requirements.txt
#
# Requires: Python 3.11+, Windows 11, NVIDIA GPU (RTX 3080)
# Avatar uses pygame with Win32 click-through transparency
# Wake word uses Whisper keyword spotting (no API keys needed)

# Environment / config
python-dotenv>=1.0.0

# Audio capture
PyAudio>=0.2.14
numpy>=1.24.0

# Speech-to-text (local, GPU-accelerated)
faster-whisper>=1.0.0

# CUDA support (required for GPU-accelerated Whisper on Windows)
nvidia-cublas-cu12
nvidia-cudnn-cu12

# AI backend ��� Anthropic Claude API
anthropic>=0.42.0

# AI backend — local model fallback (Ollama)
httpx>=0.27.0

# Text-to-speech (local Piper TTS — ONNX inference)
piper-tts>=1.4.0
pygame>=2.5.0

# Image generation (sprite creation)
Pillow>=10.0.0

# Scheduler (reminders and calendar)
APScheduler>=3.10.0
SQLAlchemy>=2.0.0
